0-spectrum workaround
now checking if the spectrum that was saved has only 0-values. If that is the case re-record the spectrum and then try again.

Just trying to save again does not work.
Getting the module data again also does not work.

The data however looks absolutely fine. I have no idea why there is this bug.